requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
feedparser==6.0.11
google-genai==0.2.2
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
